fix : 공통ui type정리
