+++ src/shared/ui/footer/Footer.tsx
