--- src/store/useCounterStore.ts
@@ -1,17 +1,0 @@
-import { create } from "zustand"
-
-interface CounterState {
-  count: number
-  increment: () => void
-  decrement: () => void
-  reset: () => void
-  setCount: (count: number) => void
-}
-
-export const useCounterStore = create<CounterState>((set) => ({
-  count: 0,
-  increment: () => set((state) => ({ count: state.count + 1 })),
-  decrement: () => set((state) => ({ count: state.count - 1 })),
-  reset: () => set({ count: 0 }),
-  setCount: (count: number) => set({ count }),
-}))
